requirements.txt
customtkinter
mysql-connector-python
psycopg2
pandas
openpyxl
sqlite3
tk